sr/mock: rename package to srfake
